Turned nav and footer into sharable components
