+++ src/app/(home)/layout.tsx
@@ -2,21 +2,9 @@ import "~/styles/globals.css";
 
 import type { Metadata } from "next";
 
-import { IconArrowUpRight, IconMenu } from "@tabler/icons-react";
 import { Montserrat, Unbounded } from "next/font/google";
-import Link from "next/link";
-import ExternalLink from "~/components/external-link";
-import NavbarItems from "~/components/navbar-items";
-import SocialLinks from "~/components/social-links";
-import { Button, buttonVariants } from "~/components/ui/button";
-import {
-  Sheet,
-  SheetContent,
-  SheetFooter,
-  SheetHeader,
-  SheetTitle,
-  SheetTrigger,
-} from "~/components/ui/sheet";
+import Footer from "~/components/footer";
+import Navbar from "~/components/navbar";
 import { Toaster } from "~/components/ui/toaster";
 import { cn } from "~/lib/cn";
 
@@ -57,80 +45,5 @@ export default function RootLayout({
         <Toaster />
       </body>
     </html>
-  );
-}
-
-function Navbar() {
-  return (
-    <header className="sticky top-0 z-10 border-b border-b-indigo-950 bg-violet-200">
-      <noscript className="inline-flex w-full justify-center bg-indigo-950 p-2 text-center text-sm font-medium text-violet-200">
-        Please enable JavaScript for the best experience.
-      </noscript>
-
-      <div className="mx-auto flex max-w-screen-2xl justify-between">
-        <Link
-          href="/"
-          className={buttonVariants({
-            size: "lg",
-            variant: "ghost",
-            class: "font-heading capitalize",
-          })}
-        >
-          Shelly Liu
-        </Link>
-
-        <NavbarItems isDrawer={false} />
-
-        <Sheet>
-          <SheetTrigger className="block sm:hidden" asChild>
-            <Button>
-              <IconMenu />
-              <span className="sr-only">Navigation menu</span>
-            </Button>
-          </SheetTrigger>
-
-          <SheetContent>
-            <SheetHeader>
-              <SheetTitle>Navigation</SheetTitle>
-            </SheetHeader>
-
-            <NavbarItems isDrawer={true} />
-
-            <SheetFooter>
-              <SocialLinks />
-            </SheetFooter>
-          </SheetContent>
-        </Sheet>
-      </div>
-    </header>
-  );
-}
-
-function Footer() {
-  const date = new Date();
-  const year = date.getFullYear();
-
-  return (
-    <footer>
-      <div className="mx-auto flex max-w-screen-md flex-col justify-between gap-6 p-6 sm:flex-row sm:items-center md:py-12">
-        <div className="space-y-1.5">
-          <p className="font-heading text-lg font-medium">
-            &copy; Shelly Liu {year}
-          </p>
-
-          <ExternalLink
-            href="https://mikeploythai.com"
-            className="gap-0.5 text-xs"
-          >
-            Built by Mike
-            <IconArrowUpRight size={14} />
-          </ExternalLink>
-        </div>
-
-        <div className="flex gap-1.5">
-          <SocialLinks />
-        </div>
-      </div>
-    </footer>
   );
 }
